Integrate DNS feature for Android
