+++ client/internal/dns/file_linux.go
@@ -1,3 +1,5 @@
+//go:build !android
+
 package dns
 
 import (
